\name{SimFlu6-package}
\alias{SimFlu6-package}
\alias{SimFlu6}
\docType{package}
\title{
  R wrapper for SimFlu6 software
}
\description{
  SimFlu6, originally written in C++, is a software for agent based model simulation for influenza transmission. This package provides seamless integration between R and C++ code for SimFlu.
}
\details{
  This package provides a wrapper R function called simflu6(). For more information, try function help (?simflu6). This package also provides a dataset with some default inputs for simflu6().
}
\author{
}
%\references{
%  This optional section can contain literature or other references for
%  background information.
%}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
%\keyword{ package }
%\seealso{
%  Optional links to other man pages
%}
\examples{
simflu6(test_data)

}
