\name{TCCestimation}
\alias{TCCestimation}
\title{Estimate VE from Traditional Case-Control Study (TCC)}
\usage{
TCCestimation(SENS, SPEC, outcomes)
}
\arguments{
\item{SENS}{sensitivity}

\item{SPEC}{specificity}

\item{outcomes}{generated outcome file}
}
\details{
1.	User Setup: Set up cohort size, sensitivity and specificity.
2.	Estimate VE
}
\value{
Mean and std of estimated VE
}
