\name{trueVE}
\alias{trueVE}
\title{Calculate the True VE}
\description{
Calculate true VE using a new modified input file
}
\usage{
trueVE(filename)
}
\arguments{
\item{filename}{Original inputfile name}
}
\details{
1.	Calculate the overall vaccine coverage based on the input parameters and generate a new modified input file. 
2.	Run SimFlu6 using the new modified input file. 
3.	Calculate true VE using the outcome file generated by SimFlu6 in step 2. 
}
\value{
True VE
}

